Add rewrite filter, calling our own URL rewriter
This is useful when you need to rewrite a URL which isn't being caught
by the full-page URL rewriter. A good example is a URL used in a block
of JavaScript code. This used to be just fine, but now that, depending
on the config, an application can work from a subdirectory, it could
happen that a URL simply doesn't work if it hasn't been rewritten.